pyproject.toml
[tool.poetry]
name = "bibparser"
version = "0.1.0"
description = ""
authors = ["Ivan Mikhailov <mikhailov.ivan.imc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
bibtexparser = "^1.4.1"
pydantic = "^2.5.3"
python-docx = "^1.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"
ruff = "^0.1.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
